Rename folder Status --> TokenStatus
