+++ simplq/src/components/pages/TokenStatus/TokenStatus.jsx
