+++ simplq/src/components/pages/TokenStatus/status.module.scss
